+++ HistoryView.m
@@ -83,7 +83,7 @@
 -(BOOL) acceptsFirstMouse:(NSEvent *)theEvent //using the tableview does not need to activate the window first
 {
   NSPoint point = [self convertPoint:[theEvent locationInWindow] fromView:nil];
-  int row = [self rowAtPoint:point];
+  NSInteger row = [self rowAtPoint:point];
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:row] byExtendingSelection:NO];
   return YES;
 }
@@ -214,7 +214,7 @@
   }
   else //if we are going down after an upwards selection, deselect last selected item
   {
-    unsigned int firstIndex = [selectedRowIndexes firstIndex];
+    NSUInteger firstIndex = [selectedRowIndexes firstIndex];
     [self deselectRow:firstIndex];
     [self scrollRowToVisible:firstIndex+1];
   }
@@ -224,7 +224,7 @@
 -(void) moveUpAndModifySelection:(id)sender
 {
   //selection to up
-  unsigned int lastSelectedRow   = [self selectedRow];
+  NSUInteger   lastSelectedRow   = [self selectedRow];
   NSIndexSet* selectedRowIndexes = [self selectedRowIndexes];
   if (lastSelectedRow == [selectedRowIndexes firstIndex]) //if the selection is going up, and up, increase it
   {
@@ -235,7 +235,7 @@
   }
   else //if we are going up after an downwards selection, deselect last selected item
   {
-    unsigned int lastIndex = [selectedRowIndexes lastIndex];
+    NSUInteger lastIndex = [selectedRowIndexes lastIndex];
     [self deselectRow:lastIndex];
     [self scrollRowToVisible:lastIndex-1];
   }
@@ -244,7 +244,7 @@
 
 -(void) moveUp:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if (selectedRow > 0)
     --selectedRow;
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
@@ -254,7 +254,7 @@
 
 -(void) moveDown:(id)sender
 {
-  int selectedRow = [self selectedRow];
+  NSInteger selectedRow = [self selectedRow];
   if ((selectedRow >= 0) && (selectedRow+1 < [self numberOfRows]))
     ++selectedRow;
   [self selectRowIndexes:[NSIndexSet indexSetWithIndex:selectedRow] byExtendingSelection:NO];
@@ -280,7 +280,7 @@
   else if ([[self window] isKeyWindow]) //if NSPointInRect(location, [clipView bounds])
   {
     location = [self convertPoint:location fromView:clipView];
-    int row = [self rowAtPoint:location];
+    NSInteger row = [self rowAtPoint:location];
     id historyItem = (row >= 0) && (row < [self numberOfRows]) ?
        [[self->historyItemsController arrangedObjects] objectAtIndex:row] : nil;
     LatexitEquation* equation = [historyItem equation];
@@ -402,7 +402,7 @@
   if (result)
   {
     NSPoint mouseLocation = [self convertPoint:[[self window] mouseLocationOutsideOfEventStream] fromView:nil];
-    int row = [self rowAtPoint:mouseLocation];
+    NSInteger row = [self rowAtPoint:mouseLocation];
     result = (row >= 0) && [self tableView:self acceptDrop:sender row:row dropOperation:NSTableViewDropOn];
     [self draggingExited:sender];//fixes a drop-ring-does-not-disappear bug
   }//end if (result)
